Access friendlist from chatsearch
Adds option to also view your friends list in the ChatSearch screen and return to ChartSearch only if you came from there instead of the information sheet.

I did wonder about extending the function CommonSetScreen() so that is could log the screen you came from and return you there next time, if a destination screen was not specified.  It might be a technically better solution than what I have added, but I'm not sure how worth while that might be unless there are more screens that become are accesible from multiple other screens.