Display Cases
have some fun locking a sub in a display case, watch them in the case like a piece of art or just as a trophy.
normal size and small size, lockable as well, similar to the cage and crate, but more visible while still contained.

if these are added please credit them under my Club name "Kirsty"