Maid Button
This PR adds a button that the user must press before maids will arrive. It can be disabled with a toggle in the preference menu. The button simply starts the timer for the maids that normally starts if the user is restrained, and does not add additional functionality to the maids.

Button icon by Nina