Sci fi items changes
Some general modifications to the Female3DCG.js for futuristic items to make them more consistent and also adding sound effects. No new features aside from edits to Audio.js to add sounds to items that didn't have any before, and edits to item scripts to allow sound effects to play. Also new sound effects themselves.

Also includes disabling random on futuristic items.

Note that sound effects only play when an action is sent to the chat--this means you need to be in a chatroom to hear it.