Maid Disabling Rank
Adds 3 'levels' for players who use the maid disabling option for long periods of time. After a long enough period of times, enables the user to disable the maids up to 3 days.

To disable maids for 6 hours, the player must have done 90 minutes of time with maids disabled first. This requirement is lower than the others, to encourage players who like being helpless for the duration of an RP, while not enabling the option for first-time players.

To disable the maids for 12 hours, the player must have spent 12 hours total with the maids off.

To disable for 72 hours, the player must have spent 72 hours already with the maids disabled.

Testing I have done so far: Sped through options to make sure progression works, make sure the new options only appear with the right amount of time remaining, made sure the new dialogue options are correct.