Immersion Settings
Moves Sensory Deprivation, Keep restraints while relogging, and disabling inspection while blind to a new Immersion menu (the icon for this one already seemed to exist).

This new menu contains an option (default: false) which, if set to true, prevents all the other items in the menu from being changed if the player is restrained (or blind).

Safeword is separate since it is a pretty important feature and remains in the General settings.