Futuristic Collar New Feature
The futuristic collar can now copy its own colors (display, band, and mesh) to other futuristic items with a new button interface.

This is a new feature, so it might make sense to keep this out of the second beta (if there is one) but it has been requested to make it easier to color a full set.