Password Lock
Adds a new lock, similar to the combination lock. However, instead of being limited to only numbers, it is limited to only 8 letter passwords made of letters, and the user can set a hint. Once a password is set, it cannot be changed.

Additionally, futuristic items will use the password entered by players and can be unlocked if the password entered in the field is the same.