Immersion Option: Stop player from leaving indoor rooms w/ hands covered
Adds a new immersion option (default: off) that allows the player to make themselves unable to leave rooms with their hands covered. 

HandsCovered is a new effect I have manually added to ItemArms and ItemLegs items and the inflatable body bag. I did not put it on boxes or cases because I want to leave open the possibility of using HandsCovered for things like lockpicking for a future escape minigame (idea being that if your hands are covered you can't hold a lockpick and can't pick locks, but if you are in a box and have a lockpick then you can free yourself from restraints you are wearing)