Ability to pick locks
Oh boy, lots of changes in this one.
1) Instead of Inspect Lock and Unlock being in the menu at the same time, there is a "Lock Menu" which contains 3 options: Inspect, Unlock, and Pick
2) Lock picking minigame was added. it is very similar to the oblivion minigame for picking locks. Challenge increases based on lock time and it is possible to extend it to make it possible to pick locks while tied up, but currently you must be able to interact with items to do it.
3) Owner, Lover, Mistress locks are not pickable. The user can try, but there will be several pins that are impossible to set no matter what, even if the user gets extremely lucky.
Combination padlocks are pickable by those with lock picking skill of 5 or higher.
There is also a new High Security padlock. I plan to update this one to allow giving keys to others so they can unlock it as well. For now, it can only be opened by the person who locked it on, and can be picked with a lock picking skill of 4 or higher.

Currently it does not require lockpicks to pick a lock. I will be working on the feature soon, with key deposits taking away lockpicks as well, and owner settings taking away the right to own lockpicks. However, there will be a character action to temporarily lend a player your lockpicks until they leave the room, and to take away a bound player's lockpicks until they leave the room. Potentially this could backfire, if lending your lockpicks takes away your own ability to pick locks.