Updates 2 Electric Boogaloo
Mainly just made it so that the Little One title is able to be used at all times since I can see many people wanting to use it instead of Baby or Diaper Lover.

Also added a second condition to Foal to allow ponies that also like ABDL to use it since Foal sounds nice in that case.