Auto-Remake rooms now remakes private, adding player as admin
Due to this:
https://cdn.discordapp.com/attachments/797897462694019093/798246477809057872/unknown.png

It became pretty clear that remaking private rooms as public will lead to a lot of spam. So instead of remaking them as public, they are instead remade as private with the caveat that the remaker is now an admin. So they can change the room parameters if they want to make it public, or choose to stay private, respecting people's wishes.

This has the side effect of making players less helpless, as a player can always make a private room public, wait for people to join, lock the room, and then F5 (since players cannot rejoin a locked room), but this was always possible if you were the room creator in the first place.