Logger Bio & Chat Messages
Lengthened the Chat Messages to 1000 characters so 4 times the original length because honestly 250 is way to short for more detailed RP.

Lengthed the max length of Bios to 10,000 characters as it is crazy hard to explain yourself in 1000 characters.  While most likely won't use it, it is nice for those that do.