Lock picking QoL + skill progress changes + pick availability
Changes the rate of lock picking skill gain so that it doesn't happen so quickly.

Also added an auto-restart if you jammed the lock.

Finally, you can only pick locks if you can walk or have access to your items. (the 'can walk' limitation is supposed to represent being able to have brought one with you before), or if someone has lent you theirs (though they need to be unbound to do so).