Virtual reality fix with light sensory deprivation
Current behavior: Being on Light sensory deprivation settings causes you to be able to see the whole chat room. There is a bug where you can't see anyone but those wearing VR headsets even if you are on Light, which goes against the purpose of Light mode.

Fix: Make it so that all players get drawn regardless of VR status when player is in Light sensory deprivation mode