Changes some futuristic stuff
Adds Futuristic Training belt, an exciting new item with a great deal of control over the wearer

Merges the futuristic chastity belts into one item to reduce clutter

Fixes an issue where the Futuristic Pleasure Panties showed both a remote icon and an extended item icon. Now it will only show the remote icon if the item is blocked.

Changed many futuristic items to use remotes instead of being AlwaysExtend, meaning that a player needs access to the controls if they are not allowed to use remotes.

Modified the futuristic vibrator to use the dark magic of regex to prevent single characters from tormenting people.